Avoid error about non-existing /etc/system-release file
When running the script with Test Kitchen, it show this error:
```
-----> Installing Ansible Omnibus
       downloading https://raw.githubusercontent.com/neillturner/omnibus-ansible/master/ansible_install.sh
         to file /tmp/ansible_install.sh
       trying wget...
       grep: /etc/system-release: No such file or directory
```
This is because of the quiet grep, but a non-existing file still triggers an error despite the quiet param. One solution would be to add the -s parameter to suppress errors, but I think the Amazon Linux check is superfluous unless I'm mistaken? Because there is already a check for /etc/system-release so in the case of Amazon Linux it will never reach that grep statement. It will however show an error on al other distro's that do not have that file.